fix bugs

+++ runner.ts
@@ -137,7 +137,7 @@ export async function run(source : string, config: Config) : Promise<[Value<Anno
     (import "js" "memory" (memory 1))
     ${wasmErrorImports}
     (func $print_num (import "imports" "print_num") (param i32) (result i32))
-    (func $print_float (import "imports" "print_float") (param i32) (result i32))
+    (func $print_float (import "imports" "print_float") (param f32) (result i32))
     (func $print_bool (import "imports" "print_bool") (param i32) (result i32))
     (func $print_none (import "imports" "print_none") (param i32) (result i32))
     (func $print_newline (import "imports" "print_newline") (param i32) (result i32))
